Inlining - Evaluator now supports run-time terms
In particular this reduces type class overhead.
NOTE: Breaks the javascript code generation because it doesn't seem to
support projection from global names, which can now arise as a result
of inlinings.

A temporary fix for this would be to update ElabDecls.hs:733 to
simple_rt ctxt t = t

which disables the simplification of run time terms.